+++ SourceCode/App/Validators/ValidatorsExtensions.cs
@@ -1,15 +1,21 @@
 ﻿using FluentValidation;
 using Microsoft.Extensions.Localization;
+using ModulesRegistry.Data;
 using ModulesRegistry.Data.Extensions;
 using ModulesRegistry.Services.Extensions;
+using System.Linq.Expressions;
+using System.Runtime.CompilerServices;
 using System.Text.RegularExpressions;
 
 namespace ModulesRegistry.Validators;
 
 public static partial class ValidatorsExtensions
 {
+
     public static IRuleBuilderOptions<T, string> MustBeCapitalizedCorrectly<T>(this IRuleBuilder<T, string> builder, IStringLocalizer localizer, bool allWordsWithInitialCapitals = true) =>
         builder.Must(value => value.IsNameCapitalizedCorrectly(allWordsWithInitialCapitals)).WithMessage(Message(localizer, allWordsWithInitialCapitals));
+
+
 
     private static string Message(IStringLocalizer localizer, bool allWordsWithInitialCapitals) =>
         allWordsWithInitialCapitals ?
@@ -71,6 +77,9 @@ public static partial class ValidatorsExtensions
 
     public static IRuleBuilderOptions<T, string> MustBeValidEmailAdresses<T>(this IRuleBuilder<T, string> builder, IStringLocalizer localizer) =>
      builder.Must(value => string.IsNullOrEmpty(value) || value.Split(';').All(email => email.IsValidEmailAddress())).WithMessage($"\"{{PropertyName}}\" {localizer["MustBeValidEmailAddresses"]}");
+
+    public static IRuleBuilderOptions<T, Station> IsRegionRequired<T>(this IRuleBuilder<T, Station> builder, IStringLocalizer localizer) =>
+        builder.Must(x => x.IsShadow || x.RegionId.HasValue).WithMessage(value => $"\"{{PropertyName}}\" {localizer["MustBeSelected"]}");
 
     private static bool IsValidEmailAddress(this string? email)
     {